+++ Ciclo 3/Asignacion de oficinas/AsigOficinas.App/AsigOficinas.App.Persistencia/Migrations/appContextModelSnapshot.cs
@@ -47,8 +47,8 @@ namespace AsigOficinas.App.Persistencia.Migrations
                     b.Property<string>("direccion")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("telefono")
-                        .HasColumnType("int");
+                    b.Property<long>("telefono")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("ubicacion")
                         .HasColumnType("nvarchar(max)");
